Refactor: LottoGameController -> LottoGameService 변경
