+++ src/main/java/lotto/Application.java
@@ -1,10 +1,10 @@
 package lotto;
 
-import lotto.controller.LottoGameController;
+import lotto.service.LottoGameService;
 
 public class Application {
     public static void main(String[] args) {
-        LottoGameController lottoGameController = new LottoGameController();
-        lottoGameController.run();
+        LottoGameService lottoGameService = new LottoGameService();
+        lottoGameService.run();
     }
 }
